UIREC-242: Support holdings-storage 6.0 in okapiInterfaces
mod-inventory-storage has new major interface versions:

instance-storage 9.0
holdings-storage 6.0
item-storage 10.0

ui-receiving uses the holdings-storage interface only, not the other two interfaces.

ui-receiving is not affected because the incompatible change is in the DELETE all APIs only that is not used by ui-receiving.

Only the okapiInterfaces need to be expanded in package.json.